// src/modules/notifications/enums/notification.codes.ts
/**
 * Notification-related error codes (NTN_xxx)
 * Range: NTN_001 - NTN_049
 */
export enum NotificationErrorCode {
  NOTIFICATION_NOT_FOUND = 'NTN_001',
  NOTIFICATION_LOG_NOT_FOUND = 'NTN_012',
  NOTIFICATION_ACCESS_DENIED = 'NTN_002',
  TEMPLATE_RENDERING_FAILED = 'NTN_003',
  INVALID_RECIPIENT = 'NTN_004',
  NOTIFICATION_ALREADY_READ = 'NTN_005',

  // Background job and system errors
  NOTIFICATION_SENDING_FAILED = 'NTN_006',
  CHANNEL_ADAPTER_FAILED = 'NTN_007',
  INVALID_CHANNEL = 'NTN_008',
  MISSING_NOTIFICATION_CONTENT = 'NTN_009',
  MISSING_TEMPLATE_VARIABLES = 'NTN_010',
  WEBHOOK_SIGNATURE_INVALID = 'NTN_011',
}